1) changing title, 2) reducing the "we-ness" of the text.
1) The original title "Open Design Definition" was very misleading since it suggested that the text the reader is about to read contains THE Open Design Definition. Actually, this text contains only a description of the Open Design Definition project. The title was changed to reflect the true content of this text.

2) A change in the beginning of the third paragraph of the "Introduction".

My initial suggestion for changing this paragraph was something like: "We hope that this project will develop..." but this was depreciated in favor of non-personalized language of traditional definitions. Rationale:

A. "We" implied a group of people which at this point in text has not yet been defined. OKFN, the public, the maintainers, the majestic mystic "we" ;)?

B. "We want to develop this missing definition through inclusive and collaborative methods rooted in public discussion" is a *contradictio in adjecto*. "we want to develop" implies that certain "we" are the developers while ... "through inclusive and collaborative methods ... public discussion" implies everyone is welcome to contribute to the development.